+++ tests/testthat/test_create_spp_aoh_data.R
@@ -87,7 +87,7 @@ test_that("simulated data (GDAL engine)", {
   skip_on_cran()
   skip_if_not_installed("rgdal")
   skip_if_not_installed("gdalUtilities")
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   # specify file path
   f <- system.file("testdata", "SIMULATED_SPECIES.zip", package = "aoh")
   elevation_data <- terra::rast(
@@ -447,7 +447,7 @@ test_that("amphibian data (IUCN format)", {
   # skip if needed
   skip_on_cran()
   skip_if_offline()
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   skip_if_iucn_key_missing()
   skip_if_iucn_api_not_available()
   skip_if_not_installed("prepr")
@@ -513,7 +513,7 @@ test_that("reptile data (IUCN format)", {
   # skip if needed
   skip_on_cran()
   skip_if_offline()
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   skip_if_iucn_key_missing()
   skip_if_iucn_api_not_available()
   skip_if_not_installed("prepr")
@@ -579,7 +579,7 @@ test_that("terrestrial mammal data (IUCN format)", {
   # skip if needed
   skip_on_cran()
   skip_if_offline()
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   skip_if_iucn_key_missing()
   skip_if_iucn_api_not_available()
   skip_if_not_installed("prepr")
@@ -645,7 +645,7 @@ test_that("bird data (BirdLife format)", {
   # skip if needed
   skip_on_cran()
   skip_if_offline()
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   skip_if_iucn_key_missing()
   skip_if_iucn_api_not_available()
   skip_if_not_installed("prepr")
@@ -712,7 +712,7 @@ test_that("bird data (alternate BirdLife format)", {
   # skip if needed
   skip_on_cran()
   skip_if_offline()
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   skip_if_iucn_key_missing()
   skip_if_iucn_api_not_available()
   skip_if_not_installed("prepr")