+++ tests/testthat/test_create_spp_frc_data.R
@@ -66,7 +66,7 @@ test_that("different engines produce same result", {
   # skip if needed
   skip_on_cran()
   skip_if_not_installed("gdalUtilities")
-  skip_if_gdal_python_not_available()
+  skip_if_gdal_calc_not_available()
   # specify file path
   f <- system.file("testdata", "SIMULATED_SPECIES.zip", package = "aoh")
   elevation_data <- terra::rast(