+++ tests/testthat/test_gdal.R
@@ -1,5 +1,5 @@
 context("gdal functions")
 
-test_that("is_gdal_python_available()", {
-  expect_is(is_gdal_python_available(), "logical")
+test_that("is_gdal_calc_available()", {
+  expect_is(is_gdal_calc_available(), "logical")
 })